Extract getHttpClient
