+++ core/api/src/commonMain/kotlin/com/nasdroid/api/v2/ApiV2Module.kt
@@ -1,6 +1,5 @@
 package com.nasdroid.api.v2
 
-import android.util.Log
 import com.nasdroid.api.v2.apikey.ApiKeyV2Api
 import com.nasdroid.api.v2.apikey.ApiKeyV2ApiImpl
 import com.nasdroid.api.v2.auth.AuthV2Api
@@ -17,16 +16,6 @@ import com.nasdroid.api.v2.reporting.ReportingV2Api
 import com.nasdroid.api.v2.reporting.ReportingV2ApiImpl
 import com.nasdroid.api.v2.system.SystemV2Api
 import com.nasdroid.api.v2.system.SystemV2ApiImpl
-import io.ktor.client.HttpClient
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.request.basicAuth
-import io.ktor.client.request.bearerAuth
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
@@ -34,43 +23,12 @@ import org.koin.dsl.module
 /**
  * A Koin module to inject the TrueNAS API V2 dependency graph.
  */
-@OptIn(ExperimentalSerializationApi::class)
 val ApiV2Module = module {
     // API state
     singleOf(::InMemoryApiStateProvider) bind ApiStateProvider::class
 
     // Ktor client
-    single {
-        val apiStateProvider: ApiStateProvider = get()
-        HttpClient {
-            // TODO if debug, BuildConfig appears to be missing
-            install(io.ktor.client.plugins.logging.Logging) {
-                level = io.ktor.client.plugins.logging.LogLevel.ALL
-                logger = object : io.ktor.client.plugins.logging.Logger {
-                    override fun log(message: String) {
-                        Log.i("Ktor", message)
-                    }
-                }
-            }
-
-            install(ContentNegotiation) {
-                json(
-                    Json {
-                        explicitNulls = false
-                    }
-                )
-            }
-            defaultRequest {
-                when (val authorization = requireNotNull(apiStateProvider.authorization)) {
-                    is Authorization.ApiKey -> bearerAuth(authorization.apiKey)
-                    is Authorization.Basic -> basicAuth(authorization.username, authorization.password)
-                }
-                val baseUrl = requireNotNull(apiStateProvider.serverAddress)
-                url(baseUrl)
-                contentType(ContentType.Application.Json)
-            }
-        }
-    }
+    singleOf(::getHttpClient)
 
     singleOf(::ApiKeyV2ApiImpl) bind ApiKeyV2Api::class
     singleOf(::AuthV2ApiImpl) bind AuthV2Api::class
